Added note entity event listener and move sendCustomerNotification
